+++ pkg/linter/test/rules/prefer_relative_imports_test.dart
@@ -16,15 +16,20 @@ void main() {
 @reflectiveTest
 class PreferRelativeImportsTest extends LintRuleTest {
   @override
-  bool get addJsPackageDep => true;
+  String get lintRule => LintNames.prefer_relative_imports;
 
   @override
-  String get lintRule => LintNames.prefer_relative_imports;
+  void setUp() {
+    newPackage('foo').addFile('lib/foo.dart', r'''
+class Foo {}
+''');
+    super.setUp();
+  }
 
   test_externalPackage() async {
     await assertNoDiagnostics(r'''
-/// This provides [JS].
-import 'package:js/js.dart';
+/// This provides [Foo].
+import 'package:foo/foo.dart';
 ''');
   }